Autoencoder/requirements.txt
#Python version 3.12.4
colorama==0.4.6
contourpy==1.2.1
cycler==0.12.1
filelock==3.15.4
fonttools==4.53.1
fsspec==2024.6.1
intel-openmp==2021.4.0
Jinja2==3.1.4
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.9.1
mkl==2021.4.0
mpmath==1.3.0
networkx==3.3
numpy==1.26.4
packaging==24.1
pillow==10.4.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
six==1.16.0
sympy==1.12.1
tbb==2021.13.0
torch==2.3.1
torchvision==0.18.1
tqdm==4.66.4
typing_extensions==4.12.2
